added insertion
Just added insertion that I forgot to write real fast.  Realised this is a real handly one to have after a discord conversation... The conversation was unrelated to insertions, but it remined me of string operations I wrote for a c program a while back and I remembered this one.  Hopefully this will work once the blitible issue is fixed and I can start testing some ot this.