Convert API to use UnsafeUtility
So I gutted out all the bad loops and replacement algorithms and built up some solid private API and examples for you to go crazy with. I also replaced the backing type with a custom struct so that we didn't have type confusion with some of the API and actual ints.

See if you can fill in the missing constructor and append operations to make the example run.